refactor: share carrier command parsing
