+++ src/infra/dispatch-wrapper-resolution.ts
@@ -1,32 +1,15 @@
 import { normalizeLowercaseStringOrEmpty } from "../shared/string-coerce.js";
-import { splitShellArgs } from "../utils/shell-argv.js";
+import {
+  envInvocationUsesModifiers,
+  parseEnvInvocationPrelude,
+  unwrapEnvInvocation,
+} from "./command-carriers.js";
 import { normalizeExecutableToken } from "./exec-wrapper-tokens.js";
+
+export { unwrapEnvInvocation } from "./command-carriers.js";
 
 export const MAX_DISPATCH_WRAPPER_DEPTH = 4;
 
-const ENV_OPTIONS_WITH_VALUE = new Set([
-  "-u",
-  "--unset",
-  "-c",
-  "--chdir",
-  "-s",
-  "--split-string",
-  "--default-signal",
-  "--ignore-signal",
-  "--block-signal",
-]);
-const ENV_INLINE_VALUE_PREFIXES = [
-  "-u",
-  "-c",
-  "-s",
-  "--unset=",
-  "--chdir=",
-  "--split-string=",
-  "--default-signal=",
-  "--ignore-signal=",
-  "--block-signal=",
-] as const;
-const ENV_FLAG_OPTIONS = new Set(["-i", "--ignore-environment", "-0", "--null"]);
 const NICE_OPTIONS_WITH_VALUE = new Set(["-n", "--adjustment", "--priority"]);
 const CAFFEINATE_OPTIONS_WITH_VALUE = new Set(["-t", "-w"]);
 const STDBUF_OPTIONS_WITH_VALUE = new Set(["-i", "--input", "-o", "--output", "-e", "--error"]);
@@ -83,19 +66,6 @@ function isKnownArchNameToken(token: string): boolean {
 
 type WrapperScanDirective = "continue" | "consume-next" | "stop" | "invalid";
 
-function isEnvAssignment(token: string): boolean {
-  return /^[A-Za-z_][A-Za-z0-9_]*=.*/.test(token);
-}
-
-function hasEnvInlineValuePrefix(lower: string): boolean {
-  for (const prefix of ENV_INLINE_VALUE_PREFIXES) {
-    if (lower.startsWith(prefix)) {
-      return true;
-    }
-  }
-  return false;
-}
-
 function scanWrapperInvocation(
   argv: string[],
   params: {
@@ -143,109 +113,6 @@ function scanWrapperInvocation(
   return argv.slice(commandIndex);
 }
 
-export function unwrapEnvInvocation(argv: string[]): string[] | null {
-  const parsed = parseEnvInvocationPrelude(argv);
-  return parsed ? (parsed.splitArgv ?? argv.slice(parsed.commandIndex)) : null;
-}
-
-type ParsedEnvInvocationPrelude = {
-  assignmentKeys: string[];
-  commandIndex: number;
-  splitArgv?: string[];
-};
-
-function splitEnvSplitStringPayload(payload: string, trailingArgv: string[]): string[] | null {
-  const splitArgv = splitShellArgs(payload);
-  return splitArgv && splitArgv.length > 0 ? [...splitArgv, ...trailingArgv] : null;
-}
-
-function parseEnvInvocationPrelude(argv: string[]): ParsedEnvInvocationPrelude | null {
-  let idx = 1;
-  let expectsOptionValue = false;
-  const assignmentKeys: string[] = [];
-  while (idx < argv.length) {
-    const token = argv[idx]?.trim() ?? "";
-    if (!token) {
-      idx += 1;
-      continue;
-    }
-    if (expectsOptionValue) {
-      expectsOptionValue = false;
-      idx += 1;
-      continue;
-    }
-    if (token === "--" || token === "-") {
-      idx += 1;
-      break;
-    }
-    if (isEnvAssignment(token)) {
-      const delimiter = token.indexOf("=");
-      if (delimiter > 0) {
-        assignmentKeys.push(token.slice(0, delimiter));
-      }
-      idx += 1;
-      continue;
-    }
-    if (!token.startsWith("-") || token === "-") {
-      break;
-    }
-    const lower = normalizeLowercaseStringOrEmpty(token);
-    const [flag] = lower.split("=", 2);
-    if (flag === "-s" || flag === "--split-string") {
-      const payload = lower.includes("=") ? token.slice(token.indexOf("=") + 1) : argv[idx + 1];
-      if (typeof payload !== "string") {
-        return null;
-      }
-      const trailingIndex = lower.includes("=") ? idx + 1 : idx + 2;
-      const splitArgv = splitEnvSplitStringPayload(payload, argv.slice(trailingIndex));
-      return splitArgv
-        ? {
-            assignmentKeys,
-            commandIndex: trailingIndex,
-            splitArgv,
-          }
-        : null;
-    }
-    if (lower.startsWith("-s") && lower.length > 2) {
-      const splitArgv = splitEnvSplitStringPayload(token.slice(2), argv.slice(idx + 1));
-      return splitArgv
-        ? {
-            assignmentKeys,
-            commandIndex: idx + 1,
-            splitArgv,
-          }
-        : null;
-    }
-    if (ENV_FLAG_OPTIONS.has(flag)) {
-      idx += 1;
-      continue;
-    }
-    if (ENV_OPTIONS_WITH_VALUE.has(flag)) {
-      if (lower.includes("=")) {
-        idx += 1;
-        continue;
-      }
-      expectsOptionValue = true;
-      idx += 1;
-      continue;
-    }
-    if (hasEnvInlineValuePrefix(lower)) {
-      idx += 1;
-      continue;
-    }
-    return null;
-  }
-
-  if (expectsOptionValue || idx >= argv.length) {
-    return null;
-  }
-
-  return {
-    assignmentKeys,
-    commandIndex: idx,
-  };
-}
-
 export function extractEnvAssignmentKeysFromDispatchWrappers(
   argv: string[],
   maxDepth = MAX_DISPATCH_WRAPPER_DEPTH,
@@ -266,50 +133,6 @@ export function extractEnvAssignmentKeysFromDispatchWrappers(
     current = unwrap.argv;
   }
   return Array.from(new Set(assignmentKeys)).toSorted((a, b) => a.localeCompare(b));
-}
-
-function envInvocationUsesModifiers(argv: string[]): boolean {
-  let idx = 1;
-  let expectsOptionValue = false;
-  while (idx < argv.length) {
-    const token = argv[idx]?.trim() ?? "";
-    if (!token) {
-      idx += 1;
-      continue;
-    }
-    if (expectsOptionValue) {
-      return true;
-    }
-    if (token === "--" || token === "-") {
-      idx += 1;
-      break;
-    }
-    if (isEnvAssignment(token)) {
-      return true;
-    }
-    if (!token.startsWith("-") || token === "-") {
-      break;
-    }
-    const lower = normalizeLowercaseStringOrEmpty(token);
-    const [flag] = lower.split("=", 2);
-    if (ENV_FLAG_OPTIONS.has(flag)) {
-      return true;
-    }
-    if (ENV_OPTIONS_WITH_VALUE.has(flag)) {
-      if (lower.includes("=")) {
-        return true;
-      }
-      expectsOptionValue = true;
-      idx += 1;
-      continue;
-    }
-    if (hasEnvInlineValuePrefix(lower)) {
-      return true;
-    }
-    return true;
-  }
-
-  return false;
 }
 
 function unwrapDashOptionInvocation(
